requirements/requirements.txt
click == 6.2
semidbm == 0.5.1
six == 1.10.0
python-RIBBIT == 1.0.2.dev0
